ops/skuld-cli/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "skuld-cli"
version = "0.1.0"
description = "CLI for deploying and managing SKULD environments via GitHub Actions"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
skuld = "skuld_cli.cli:main"
